refactor button component
