--- src/components/Buttons/Buttons.scss
@@ -1,18 +1,0 @@
-.transparent-btn {
-  text-decoration: none;
-  display: inline-flex;
-  align-items: center;
-  background: transparent;
-  padding: 0.75rem 1.25rem;
-  color: white;
-  font-size: var(--font-sm);
-  border: 1px solid var(--dark-blue-border);
-  border-radius: var(--border-radius-sm);
-  cursor: pointer;
-  &:hover {
-    background: var(--dark-blue-hover);
-  }
-  &:active {
-    background: var(--dark-blue-active);
-  }
-}
