--- src/components/Buttons/TransparentButton.tsx
@@ -1,5 +1,0 @@
-function TransparentButton({ onClick, ...rest }) {
-  return <button onClick={onClick} className="transparent-btn" {...rest}></button>;
-}
-
-export default TransparentButton;
